refactor: Move EEPROM layout to hal and update test structure
