+++ platformio.ini
@@ -157,7 +157,7 @@ build_src_flags =
 	-Werror
 	-Wpedantic
 	-Wshadow
-test_lib_deps =
-	ArduinoFake@^0.4.0
 extra_scripts = pre:scripts/test-coverage.py
-test_filter = test_core
+test_filter =
+	test_core
+	test_hal
